update tests and api spec
